Add files via upload
Recently, I was asked about Value at Risk (VaR) and thought of implementing an example in Python. The code here illustrates stressed VaR, which is a 10 day (holding period) unweighted (all points in time series treated equally) VaR for a 99% confidence interval.

The time series used is for Turkish Lira FX rate during the 2008 period.

It’s a fairly simple exercise.

However, the points to ponder are:

What if a timeseries in not available? E.g. EUR in 2000
What if the timeseries is discontinuous? E.g. FX peg break.
How to consider restructurings, takeover and succession events?